#57 Refactor: Utils 내 스크립트 파일 세분화 및 변동 사항 반영
