+++ wtnt/core/utils/s3.py
@@ -1,0 +1,18 @@
+import boto3
+
+import wtnt.settings as settings
+
+
+class S3Utils:
+    @staticmethod
+    def upload_s3(name, image):
+        s3_client = boto3.client(
+            "s3",
+            aws_access_key_id=settings.AWS_CLIENT_ID,
+            aws_secret_access_key=settings.AWS_CLIENT_SECRET,
+            region_name=settings.AWS_REGION,
+        )
+        root = "team/" + name + "/image.jpg"
+        s3_client.upload_fileobj(image, settings.BUCKET_NAME, root)
+
+        return f"https://{settings.BUCKET_NAME}.s3.{settings.AWS_REGION}.amazonaws.com/{root}"
