+++ wtnt/core/utils/team.py
@@ -1,12 +1,4 @@
-import boto3
-
-from django_redis import get_redis_connection
-
 from team.models import Likes
-import wtnt.settings as settings
-
-
-client = get_redis_connection()
 
 
 class TeamResponse:
@@ -96,24 +88,3 @@ class LikeResponse:
     @staticmethod
     def make_data(like_num, is_like, version):
         return {"like": {"like_count": like_num, "is_like": is_like}, "version": version}
-
-
-class S3Utils:
-    @staticmethod
-    def upload_s3(name, image):
-        s3_client = boto3.client(
-            "s3",
-            aws_access_key_id=settings.AWS_CLIENT_ID,
-            aws_secret_access_key=settings.AWS_CLIENT_SECRET,
-            region_name=settings.AWS_REGION,
-        )
-        root = "team/" + name + "/image.jpg"
-        s3_client.upload_fileobj(image, settings.BUCKET_NAME, root)
-
-        return f"https://{settings.BUCKET_NAME}.s3.{settings.AWS_REGION}.amazonaws.com/{root}"
-
-
-class RedisTeamUtils:
-    @staticmethod
-    def sadd_view_client(team_id, user_id, adress):
-        return client.sadd(f"views:{team_id}", f"{user_id}_{adress}")
