+++ wtnt/user/auth/service.py
@@ -3,7 +3,8 @@ from django.contrib.auth import get_user_model
 from rest_framework_simplejwt.tokens import AccessToken
 from rest_framework_simplejwt.exceptions import TokenError, InvalidToken
 
-from core.utils.auth import get_user_info, RedisAuthUtils
+from core.utils.auth import get_user_info
+from core.utils.redis import RedisUtils
 from core.service import BaseService
 from core.exceptions import CodeNotMatchError, RefreshTokenExpiredError, CeleryTaskError
 from user.tasks import send_email
@@ -27,7 +28,7 @@ class AuthService(BaseService):
         user_id = response_data["user"]["pk"]
 
         if refresh_token:
-            RedisAuthUtils.set_refresh_token_in_cache(user_id, refresh_token)
+            RedisUtils.set_refresh_token_in_cache(user_id, refresh_token)
             response_data.pop("refresh")
 
         _, email = response_data["user"]["email"].split("@")
@@ -48,14 +49,14 @@ class RegisterService(BaseService):
     def finish_register_by_user_input(self):
         code = self.request.data.get("code")
         email = self.request.data.get("email")
-        if code != RedisAuthUtils.get_code_in_redis_from_email():
+        if code != RedisUtils.get_code_in_redis_from_email():
             raise CodeNotMatchError()
 
         extra_data = SocialAccount.objects.get(user_id=self.request.user.id).extra_data
         user = User.objects.get(id=self.request.user.id)
         user.finish_register(extra_data=extra_data, request_data=self.request.data)
 
-        RedisAuthUtils.delete_code_in_redis_from_email(email)
+        RedisUtils.delete_code_in_redis_from_email(email)
 
         return user
 
@@ -65,7 +66,7 @@ class RefreshService(BaseService):
         _, access_token = self.request.META.get("HTTP_AUTHORIZATION").split(" ")
         user_id = AccessToken(access_token, verify=False).payload.get("user_id")
 
-        refresh_token = RedisAuthUtils.get_refresh_token_in_cache(user_id)
+        refresh_token = RedisUtils.get_refresh_token_in_cache(user_id)
         if not refresh_token:
             raise RefreshTokenExpiredError()
 
@@ -95,8 +96,8 @@ class EmailVerifyService(BaseService):
         code = self.request.data.get("code")
         email = self.request.data.get("email")
 
-        if code == RedisAuthUtils.get_code_in_redis_from_email(email):
-            RedisAuthUtils.set_code_in_redis_from_email(email, code)
+        if code == RedisUtils.get_code_in_redis_from_email(email):
+            RedisUtils.set_code_in_redis_from_email(email, code)
             return code
 
         raise CodeNotMatchError()
